moving dialog to docmanager
