+++ packages/docmanager-extension/package.json
@@ -33,12 +33,11 @@
     "@jupyterlab/application": "^0.15.2",
     "@jupyterlab/apputils": "^0.15.2",
     "@jupyterlab/coreutils": "^1.0.4",
-    "@jupyterlab/docmanager": "^0.15.2",
+    "@jupyterlab/docmanager": "file:../docmanager",
     "@jupyterlab/docregistry": "^0.15.2",
     "@jupyterlab/mainmenu": "^0.4.2",
     "@jupyterlab/services": "^1.1.2",
-    "@phosphor/disposable": "^1.1.2",
-    "@phosphor/widgets": "^1.5.0"
+    "@phosphor/disposable": "^1.1.2"
   },
   "devDependencies": {
     "rimraf": "~2.6.2",
